+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.13.0
+
+### New Features
+
+* Adds a new return type for `IOClient` that includes the ability to detach a 
+  socket. To use or reference `IOStreamedResponse` directly, you will need to 
+  import the new `package:http/io_streamed_response.dart` import. Then cast 
+  a `StreamedResponse` as an `IOStreamedReesponse` to call the `detachSocket()`
+  method. 
+
 ## 0.12.1-dev
 
 * Remove dependency on `package:async`.